requirements.txt
anthropic>=0.40.0
finnhub-python>=2.4.20
edgartools>=3.0.0
yfinance>=0.2.50
feedparser>=6.0.11
praw>=7.8.1
fastapi>=0.100.0
uvicorn>=0.30.0
pandas>=2.2.0
plotly>=5.24.0
python-dotenv>=1.0.1
apscheduler>=3.10.4
requests>=2.32.0
